Refactor v3 orders API into modular components
