+++ src/api/v3/orders/mod.rs
@@ -1,0 +1,12 @@
+pub mod adapters;
+pub mod errors;
+pub mod handlers;
+pub mod mappers;
+pub mod types;
+pub mod validators;
+
+#[allow(unused_imports)]
+pub use handlers::{
+    __path_cancel_order, __path_get_order, __path_my_trades, __path_new_order, __path_open_orders,
+    cancel_order, get_order, my_trades, new_order, open_orders, router,
+};
